+++ wiki/Changelog.md
@@ -1,3 +1,7 @@
+Terminal features
+  * Report DECRQSS DECSCL conformance level as VT500 (7-bit controls).
+  * Rectangular checksum DECRQCRA.
+
 Window handling
   * Reenable left scrollbar.